chore: working-state edits from device bring-up
User's Android Studio changes made during live sync bring-up, all
verified working on-device (sync succeeds end-to-end):

- DecaySettingsEntity Float -> Double, with rationale: bit-identical
  parity with the server's Python 64-bit floats (§1.1/§4.1). Room
  maps Double->REAL same as Float, so NO schema change / no migration
  / exported v4 schema still valid.
- CalculateReadinessUseCase: drop now-redundant .toDouble() calls
  (weights are already Double). Algorithm logic unchanged — parity
  fixtures still pass.
- ReadinessRepositoryImpl relocated to data.repository (proper
  package/path); old domain/.../ReadinessRepositoryImp.kt left as a
  pointer stub. Sync-metadata stamping (§4.3/§1.5) preserved.
- AndroidManifest: reference @xml/network_security_config (file was
  committed earlier; this wires it — localhost cleartext only, no
  effect on the Tailscale HTTPS sync).
- SettingsScreen: Float/Double slider adaptation + server-sync entry.
- ServerApi: comment touch-up.
- gradle.properties: unpin machine-specific org.gradle.java.home.

verify-no-secrets: CLEAN. No Room version bump. Build + live device
sync verified by user.

+++ app/src/main/java/com/trybsportowy/data/local/DecaySettingEntity.kt
@@ -3,12 +3,18 @@ package com.trybsportowy.data.local
 import androidx.room.Entity
 import androidx.room.PrimaryKey
 
+/**
+ * Decay weights for the readiness algorithm (CLAUDE.md §4.1.1).
+ *
+ * Uses [Double] to ensure bit-identical results with the server's Python implementation
+ * (which uses 64-bit floats).
+ */
 @Entity(tableName = "decay_settings")
 data class DecaySettingsEntity(
     @PrimaryKey val id: Int = 1,
-    val weightToday: Float = 1.0f,
-    val weightYesterday: Float = 0.75f,
-    val weightYesterdaySleep: Float = 0.8f,
-    val weightTwoDaysAgo: Float = 0.5f,
-    val weightThreeDaysAgo: Float = 0.2f
+    val weightToday: Double = 1.0,
+    val weightYesterday: Double = 0.75,
+    val weightYesterdaySleep: Double = 0.8,
+    val weightTwoDaysAgo: Double = 0.5,
+    val weightThreeDaysAgo: Double = 0.2
 )
